Problems with insert

+++ experiment.py
@@ -21,8 +21,7 @@ class Task(dj.Lookup):
 class Session(dj.Manual):
     definition = """
     -> lab.Animal
-    #session : smallint 		# session number (change to int for now)
-    session : int 		# session number
+    session : smallint 		# session number (change to int for now)
     ---
     session_date  : date
     -> lab.Person
